% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompStats.R
\name{CompStats}
\alias{CompStats}
\title{CompStats}
\usage{
CompStats(inputs, SimResults)
}
\arguments{
\item{inputs}{Inputs from .rav file}

\item{SimResults}{A list from \link{RunSims2R} or \link{RunSims2C} with totEsc and totAEQMort for each ER/Pop value, each sim rep, and each year.}
}
\value{
SummaryStats and staticvars list
}
\description{
Returns the statistics (calculated values) needed to produce the summary output files
}
\details{
This function similar but not identical to the original VB function in VRAP
}
