% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInput2.R
\name{GetInput2}
\alias{GetInput2}
\title{GetInput for VRAP > 2.0}
\usage{
GetInput2(InFile)
}
\arguments{
\item{InFile}{the name of the .rav file}
}
\value{
Returns the list of all inputs
}
\description{
Read in a .rav file from the original VRAP vrs > 2.0 and assign all the variables
}
\details{
VRAP 2.0 uses the same .rav file as VRAP 1.0 but some input lines are ignored.  VRAP 2.0 will also check for a tag `VRAP2` at the start of the .rav file.  If it is present, `GetInput2` knows the rav file is properly adjusted for VRAP  2.0 (VRAPS).  If the tag is not present, the .rav file is assumed to be in VRAP 1.0 form and proceeds unless option that is illegal in VRAP 2.0 is encountered (like covariates).  See the demo rav file in the `inst` folder (see examples).
}
\examples{
\dontrun{
fpath <- system.file("VRAP", "demofiles/Demo-ER.rav", package="VRAPS")
file.show(fpath)
}
}
