% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsBackwardCompat.R
\name{InputsBackwardCompat}
\alias{InputsBackwardCompat}
\title{Backwards Compatibility for inputs List}
\usage{
InputsBackwardCompat(inputs)
}
\arguments{
\item{inputs}{A list of the necessary input values (can be taken from .rav file).}
}
\value{
Returns the list of all inputs with added values so that list is compatible with VRAP 1.0 functions
}
\description{
Ensure that inputs list can be used in VRAP 1.0 functions
}
\details{
The VRAP 2.0 inputs list is very similar to VRAP 1.0 but has a few differences related to how ER and Pop steps are named.  This creates entries in inputs that has the same names used in VRAP 1.0 so that the inputs list in VRAP 2.0 can be passed to VRAP 1.0 functions.
}
