% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSims2.R
\name{RunSims2}
\alias{RunSims2}
\title{RunSims2}
\usage{
RunSims2(inputs)
}
\arguments{
\item{inputs}{Inputs from .rav file}

\item{silent}{Whether to show progress bar}
}
\value{
list with inputs, SummaryStats, staticvars, comp.time.
}
\description{
RunSims2 takes the input list, runs the VRAP simulations, and returns the summary statistics used by VRAP 1.0
}
\details{
This function is to produce VRAP 1.0 output stats using VRAP 1.0 functions.
}
