% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSims2C.R
\name{RunSims2C}
\alias{RunSims2C}
\title{Run VRAP 2.0 Simulations in C++}
\usage{
RunSims2C(inputs, silent = TRUE)
}
\arguments{
\item{silent}{Whether print progress as the current ER value.}

\item{input}{A list of the necessary input values (can be taken from .rav file).}
}
\value{
A list with the input and the 3D array of total escapment values.
}
\description{
Run the VRAP2 simulations in C++ over a specified range of escapement rates (ERs).
}
\details{
Calls the C++ function `simFish` to run the VRAP simulations
 and returns a 3D array of the total escapement
 at each exploitation rate (ER) for NRuns over NYears.
}
