% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSims2R.R
\name{RunSims2R}
\alias{RunSims2R}
\title{Run VRAP 2.0 Simulations in native R}
\usage{
RunSims2R(inputs, silent = TRUE)
}
\arguments{
\item{inputs}{A list of the necessary input values (can be taken from .rav file).}

\item{silent}{Whether print progress as the current ER value.}
}
\value{
A list with the inputs and the 3D array of total escapment values.
}
\description{
Run the VRAP2 simulations in native R over a specified range of escapement rates (ERs).
}
\details{
Runs the simulations and returns an 3D array of the total escapement
at each exploitation rate (ER) for NRuns over NYears.
}
