% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteRavFile.R
\name{WriteRavFile}
\alias{WriteRavFile}
\title{Write Rav File from VRAP 2.0 inputs list}
\usage{
WriteRavFile(inputs, FileName = "tmp.rav")
}
\arguments{
\item{inputs}{A list of the necessary input values for \link{RunSims2R} and \link{RunSims2C}.}

\item{FileName}{Name of the .rav file that data will be written to.}
}
\value{
Nothing. The data is written to ravFileName.
}
\description{
Takes the input list and write a .rav file.
}
\details{
Takes the inputs list and write a .rav file that can be input into the VRAP 1.0 function \link[VRAP]{Main}.
}
