% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteRavFile2.R
\name{WriteRavFile2}
\alias{WriteRavFile2}
\title{Write Rav File from input list}
\usage{
WriteRavFile2(inputs, ravFileName = "tmp.rav", VRAPvrs = 1)
}
\arguments{
\item{ravFileName}{Name of the .rav file that data will be written to.}

\item{input}{A list of the necessary input values for \link{RunSims2R} and \link{RunSims2C}.}
}
\value{
Nothing. The data is written to ravFileName.
}
\description{
Takes the input list and write a .rav file.
}
\details{
Takes the input list and write a .rav file that can be input into the VRAP function \link[VRAP]{Main}.
}
