\name{VRAPCpp-package}
\alias{VRAPCpp-package}
\alias{VRAPCpp}
\docType{package}
\title{
  VRAP with C++ for RER and Viability Computations
}
\description{
  This is a rewrite of the R version of the VRAP program.  The original VRAP R package was a one-to-one translation from the original Visual Basic code.  VRAPCpp is the same equations, but completely re-written by Martin Liermann to be more efficient.  VRAPCpp does not have all the functionality of VRAP. Many of the rav options in VRAP were not used and are removed.
  }
\details{
  Call `vignette(package = "VRAPCpp")` to see all the package documentation and examples.
}
\author{
  Martin Liermann and Elizabeth Holmes
  
  Maintainer: Elizabeth Holmes <eli.holmes@noaa.gov>
}
\keyword{ package }
